requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
flask-socketio==5.1.1
gunicorn==20.1.0
requests>=2.31
Werkzeug==2.0.3
pytest==6.2.4
google-cloud-storage==1.42.3
google-cloud-pubsub==2.7.0
google-cloud-monitoring==2.7.0